+++ package/avigilon/alta/access/src/CollectorAvigilonAltaAccessImpl.ts
@@ -1,3 +1,10 @@
+import {
+  AvigilonAltaEntry,
+  AvigilonAltaGroup,
+  AvigilonAltaSite,
+  AvigilonAltaUser,
+  AvigilonAltaZone
+} from '@auditlogic/schema-avigilon-alta-access-ts';
 import { ConnectionMetadata } from '@auditmation/hub-core';
 import {
   InvalidStateError,
@@ -7,8 +14,9 @@ import {
 import { Batch } from '@auditmation/util-collector-utils/dist/src';
 import { Parameters } from 'generated/model';
 import { injectable } from 'inversify';
+import PromisePool from '@supercharge/promise-pool';
 import { BaseClient } from '../generated/BaseClient';
-import { mapGroup, mapUser } from './Mappers';
+import { mapEntry, mapGroup, mapSite, mapUser, mapZone } from './Mappers';
 
 @injectable()
 export class CollectorAvigilonAltaAccessImpl extends BaseClient {
@@ -28,8 +36,11 @@ export class CollectorAvigilonAltaAccessImpl extends BaseClient {
   }
 
   private classes = {
-    user: 'AvigilonAltaUser',
-    group: 'AvigilonAltaGroup',
+    user: AvigilonAltaUser,
+    group: AvigilonAltaGroup,
+    site: AvigilonAltaSite,
+    zone: AvigilonAltaZone,
+    entry: AvigilonAltaEntry,
   };
 
   private async init() {
@@ -42,19 +53,34 @@ export class CollectorAvigilonAltaAccessImpl extends BaseClient {
   }
 
   private async initBatchForClass<T extends Record<string, any>>(
-    className: string,
+    batchItemType: new (...args) => T,
     groupId?: string
   ): Promise<Batch<T>> {
     const batch: Batch<T> = new Batch<T>(
-      className,
+      batchItemType.name,
       this.platform,
       this.logger,
       this.jobId,
-      this.metadata?.tags || [],
+      this.metadata?.tags,
       groupId
     );
     await batch.getId();
     return batch;
+  }
+
+  private async identifyOrganization(): Promise<void> {
+    if (this.orgId) {
+      this.logger.info(`Organization ID already set to ${this.orgId}`);
+      return;
+    }
+    this.logger.info('Identifying organization ID by token scope');
+    const tokenInfo = await this.access.getAuthApi().getTokenProperties();
+    if (tokenInfo.organizationId) {
+      this.orgId = `${tokenInfo.organizationId}`;
+      this.logger.info(`Identified organization ID: ${this.orgId}`);
+    } else {
+      throw new InvalidStateError('Organization ID is not set and cannot be identified from token');
+    }
   }
 
   private async loadUsers(): Promise<void> {
@@ -66,8 +92,7 @@ export class CollectorAvigilonAltaAccessImpl extends BaseClient {
 
     const usersPr = await this.access.getUserApi().list(this.orgId);
     await usersPr.forEach(async (user) => {
-      const userInfo = await this.access.getUserApi().get(this.orgId!, user.id);
-      await userBatch.add(mapUser(userInfo));
+      await userBatch.add(mapUser(user));
     }, 3, this.previewCount);
     await userBatch.end();
   }
@@ -80,22 +105,80 @@ export class CollectorAvigilonAltaAccessImpl extends BaseClient {
     const groupBatch = await this.initBatchForClass(this.classes.group, this.orgId);
     const groupsPr = await this.access.getGroupApi().list(this.orgId);
     await groupsPr.forEach(async (group) => {
-      const groupInfo = await this.access.getGroupApi().get(this.orgId!, group.id);
       const groupMembersPr = await this.access.getGroupApi().listUsers(this.orgId!, group.id);
       const membersIds: string[] = [];
       await groupMembersPr.forEach(async (member) => {
         membersIds.push(`${member.id}`);
       });
-      await groupBatch.add(mapGroup(groupInfo, membersIds));
+      await groupBatch.add(mapGroup(group, membersIds));
     }, 3, this.previewCount);
     await groupBatch.end();
+  }
+
+  private async loadSites(): Promise<void> {
+    if (!this.orgId) {
+      throw new InvalidStateError('Organization ID is not set');
+    }
+
+    const siteBatch = await this.initBatchForClass(this.classes.site, this.orgId);
+    const sites = await this.access.getSiteApi().list(this.orgId);
+    await PromisePool.for(sites)
+      .withConcurrency(3)
+      .handleError(async (error, site) => {
+        await siteBatch.error(`Error processing site ${site.id}`, error);
+      })
+      .process(async (site) => {
+        await siteBatch.add(mapSite(site));
+      });
+    await siteBatch.end();
+  }
+
+  private async loadZones(): Promise<void> {
+    if (!this.orgId) {
+      throw new InvalidStateError('Organization ID is not set');
+    }
+
+    const zoneBatch = await this.initBatchForClass(this.classes.zone, this.orgId);
+    const zones = await this.access.getZoneApi().list(this.orgId);
+    await PromisePool.for(zones)
+      .withConcurrency(3)
+      .handleError(async (error, zone) => {
+        await zoneBatch.error(`Error processing zone ${zone.id}`, error);
+      })
+      .process(async (zone) => {
+        await zoneBatch.add(mapZone(zone));
+      });
+    await zoneBatch.end();
+  }
+
+  private async loadEntries(): Promise<void> {
+    if (!this.orgId) {
+      throw new InvalidStateError('Organization ID is not set');
+    }
+
+    const entryBatch = await this.initBatchForClass(this.classes.entry, this.orgId);
+    const entries = await this.access.getEntryApi().list(this.orgId);
+    await PromisePool.for(entries)
+      .withConcurrency(3)
+      .handleError(async (error, entry) => {
+        await entryBatch.error(`Error processing entry ${entry.id}`, error);
+      })
+      .process(async (entry) => {
+        await entryBatch.add(mapEntry(entry));
+      });
+    await entryBatch.end();
   }
 
   public async run(parameters?: Parameters): Promise<any> {
     await this.init();
     this.orgId = parameters?.organizationId;
 
+    await this.identifyOrganization();
+
     await this.loadUsers();
     await this.loadGroups();
+    await this.loadSites();
+    await this.loadZones();
+    await this.loadEntries();
   }
 }