+++ package/avigilon/alta/access/src/Mappers.ts
@@ -1,5 +1,6 @@
 import * as s from '@auditlogic/schema-avigilon-alta-access-ts';
 import { PrincipalType } from '@auditlogic/schema-avigilon-alta-access-ts';
+import { GeoCountry, GeoCountryDef, PhoneNumber } from '@auditmation/types-core-js';
 import * as m from '@zerobias-org/module-avigilon-alta-access';
 
 function toUserStatus(raw?: m.UserInfo.StatusEnumDef): s.AccountStatus | undefined {
@@ -15,17 +16,15 @@ function toUserStatus(raw?: m.UserInfo.StatusEnumDef): s.AccountStatus | undefin
   }
 }
 
-export function mapUser(raw: m.UserInfo): s.Account {
+export function mapUser(raw: m.User): s.Account {
   const output: s.Account = {
     id: `${raw.id}`,
-    name: `${raw.firstName ?? ''} ${raw.lastName ?? ''}`.trim(),
-    email: raw.email,
-    identity: `${raw.email}`,
-    login: `${raw.email}`,
-    person: `${raw.email}`,
+    name: `${raw.identity?.firstName ?? ''} ${raw.identity?.lastName ?? ''}`.trim() || `User ${raw.id}`,
+    email: raw.identity?.email,
+    identity: `${raw.identity?.email}`,
+    login: `${raw.identity?.email}`,
+    person: `${raw.identity?.email}`,
     status: toUserStatus(raw.status),
-    app: raw.organizationId ? `${raw.organizationId}` : undefined,
-    icon: raw.avatarUrl,
     principalType: PrincipalType.USER,
   };
   Object.assign(
@@ -38,12 +37,11 @@ export function mapUser(raw: m.UserInfo): s.Account {
   return output;
 }
 
-export function mapGroup(raw: m.GroupInfo, memberIds: string[]): s.Group {
+export function mapGroup(raw: m.Group, memberIds: string[]): s.Group {
   const output: s.Group = {
     id: `${raw.id}`,
-    name: raw.name,
+    name: raw.name || `Group ${raw.id}`,
     description: raw.description,
-    groups: raw.parentGroupId ? `${raw.parentGroupId}` : undefined,
     principalType: PrincipalType.GROUP,
     members: memberIds.map((id) => `${id}`),
   };
@@ -54,5 +52,78 @@ export function mapGroup(raw: m.GroupInfo, memberIds: string[]): s.Group {
       dateLastModified: raw.updatedAt?.toISOString().split('T')[0],
     }
   );
+  return output;
+}
+
+export function mapSite(raw: m.Site): s.AvigilonAltaSite {
+  let country: GeoCountryDef | undefined;
+  try {
+    country = GeoCountry.from('{raw.country}');
+  } catch {
+    // ignore invalid country codes
+  }
+
+  let phoneNo: PhoneNumber | undefined;
+  try {
+    phoneNo = new PhoneNumber(raw.phone || '');
+  } catch {
+    // ignore invalid phone numbers
+  }
+
+  const output: s.AvigilonAltaSite = {
+    id: `${raw.id}`,
+    name: raw.name || `Site ${raw.id}`,
+    locations: [{
+      address: {
+        addressLine: [`${raw.address}`, `${raw.address2}`].filter(Boolean).join(', '),
+        locality: raw.city,
+        postalCode: raw.zip,
+        country,
+      },
+      phoneNumber: phoneNo,
+    }],
+  };
+
+  Object.assign(
+    output,
+    {
+      dateCreated: raw.createdAt?.toISOString().split('T')[0],
+      dateLastModified: raw.updatedAt?.toISOString().split('T')[0],
+    }
+  );
+  return output;
+}
+
+export function mapZone(raw: m.Zone): s.AvigilonAltaZone {
+  const output: s.AvigilonAltaZone = {
+    id: `${raw.id}`,
+    name: raw.name || `Zone ${raw.id}`,
+    description: raw.description,
+    site: `${raw.site?.id}`,
+  };
+
+  Object.assign(output, {
+    dateCreated: raw.createdAt?.toISOString().split('T')[0],
+    dateLastModified: raw.updatedAt?.toISOString().split('T')[0],
+  });
+
+  return output;
+}
+
+export function mapEntry(raw: m.EntryDetails): s.AvigilonAltaEntry {
+  const state = s.PhysicalEntry_status[raw.entryState?.name.toUpperCase() || 'UNKNOWN'];
+  const output: s.AvigilonAltaEntry = {
+    id: `${raw.id}`,
+    name: raw.name || `Entry ${raw.id}`,
+    note: raw.notes,
+    zone: raw.zone?.id ? `${raw.zone.id}` : undefined,
+    aliases: raw.externalUuid,
+    state,
+  };
+
+  Object.assign(output, {
+    dateCreated: raw.createdAt?.toISOString().split('T')[0],
+    dateLastModified: raw.updatedAt?.toISOString().split('T')[0],
+  });
   return output;
 }